+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/javadoc/JavadocParagraphCheckTest.java
@@ -47,7 +47,8 @@ public class JavadocParagraphCheckTest extends BaseCheckTestSupport {
     @Override
     protected String getPath(String filename) throws IOException {
         return super.getPath("checks" + File.separator
-                + "javadoc" + File.separator + filename);
+                + "javadoc" + File.separator
+                + "javadocparagraph" + File.separator + filename);
     }
 
     @Test
@@ -62,7 +63,7 @@ public class JavadocParagraphCheckTest extends BaseCheckTestSupport {
     public void testCorrect() throws Exception {
         final String[] expected = CommonUtils.EMPTY_STRING_ARRAY;
 
-        verify(checkConfig, getPath("InputCorrectJavaDocParagraph.java"), expected);
+        verify(checkConfig, getPath("InputJavaDocParagraphCorrect.java"), expected);
     }
 
     @Test
@@ -105,7 +106,7 @@ public class JavadocParagraphCheckTest extends BaseCheckTestSupport {
             "81: " + getCheckMessage(MSG_TAG_AFTER),
             "82: " + getCheckMessage(MSG_TAG_AFTER),
         };
-        verify(checkConfig, getPath("InputIncorrectJavaDocParagraph.java"), expected);
+        verify(checkConfig, getPath("InputJavaDocParagraphIncorrect.java"), expected);
     }
 
     @Test
@@ -132,6 +133,6 @@ public class JavadocParagraphCheckTest extends BaseCheckTestSupport {
             "81: " + getCheckMessage(MSG_TAG_AFTER),
             "82: " + getCheckMessage(MSG_TAG_AFTER),
         };
-        verify(checkConfig, getPath("InputIncorrectJavaDocParagraph.java"), expected);
+        verify(checkConfig, getPath("InputJavaDocParagraphIncorrect.java"), expected);
     }
 }
